+++ src/monitor/worker.rs
@@ -21,7 +21,7 @@
 // each t seconds:
 //              by using channel to blockchain dispatcher
 //              request dynamic infos to core via GetCoreStatsRequest to blockchain
-//              recive infos via GetCoreStatsresponse
+//              receive infos via GetCoreStatsresponse
 //              send infos to all the stations
 use ascii_table::{Align, AsciiTable};
 use isahc::{Request, RequestExt};
@@ -45,7 +45,7 @@ pub struct UnconfirmedPool {
 }
 
 #[derive(Serialize)]
-/// structure that holds the last block recieved by the node and its hash
+/// structure that holds the last block received by the node and its hash
 pub struct LastBlock {
     block: Block,
     hash: Hash,
@@ -63,8 +63,6 @@ pub struct P2pInfo {
 #[derive(Serialize)]
 pub struct NetworkConfig {
     pub name: String,
-    // it should be the bootstrap hash
-    //network_id: Hash, todo!()
     pub block_threshold: usize,
     pub block_timeout: u16,
 }
@@ -94,19 +92,17 @@ pub struct Status {
     pub core_version: String,
     /// last node's block
     pub last_block: Option<LastBlock>,
-    /// structure that holds some infomration about the unconfirmed tx queue
+    /// structure that holds some information about the unconfirmed tx queue
     pub unconfirmed_pool: Option<UnconfirmedPool>,
-    /// infos reuarding the p2p config
+    /// infos regarding the p2p config
     pub p2p_info: P2pInfo,
     /// seed
     pub seed: u64,
-    // TODO
-    //rcv_message_in_window: T,
 }
 
-// due to server interaction the Monitor server
-// structure needs this names as field
-/// It holds the node informations
+/// Due to server interaction the Monitor server
+/// Structure needs this names as field
+/// It holds the node information
 #[derive(Serialize)]
 #[allow(non_snake_case)]
 pub struct MonitorConfig {
@@ -117,11 +113,16 @@ pub struct MonitorConfig {
 pub struct MonitorWorker {
     config: MonitorConfig,
     bc_chan: BlockRequestSender,
+    offline: bool,
 }
 
 impl MonitorWorker {
-    pub fn new(config: MonitorConfig, bc_chan: BlockRequestSender) -> Self {
-        MonitorWorker { config, bc_chan }
+    pub fn new(config: MonitorConfig, bc_chan: BlockRequestSender, offline: bool) -> Self {
+        MonitorWorker {
+            config,
+            bc_chan,
+            offline,
+        }
     }
 
     /// Updates node status
@@ -134,7 +135,7 @@ impl MonitorWorker {
             self.config.data.last_block = Some(last_block);
         }
 
-        // retireve seed
+        // Retrieve the seed
         let request = Message::GetSeedRequest;
         let rx_chan = match self.bc_chan.send_sync(request) {
             Ok(rx_chan) => rx_chan,
@@ -246,7 +247,6 @@ impl MonitorWorker {
 
         let network_data: Vec<Vec<&dyn Display>> = vec![
             vec![&"network name", &self.config.data.nw_config.name],
-            //vec![&"network id", &network_id], todo!()
             vec![
                 &"block threshold",
                 &self.config.data.nw_config.block_threshold,
@@ -387,18 +387,20 @@ impl MonitorWorker {
 
             match rx_chan.recv_sync() {
                 Ok(Message::GetCoreStatsResponse(info)) => {
-                    if info.1 > 0 {
-                        let unconfirmed_pool = Some(UnconfirmedPool {
-                            hash: info.0,
-                            size: info.1,
-                        });
-                        self.update(info.2, unconfirmed_pool);
-                    } else {
-                        self.update(info.2, None)
-                    }
+                    if !self.offline {
+                        if info.1 > 0 {
+                            let unconfirmed_pool = Some(UnconfirmedPool {
+                                hash: info.0,
+                                size: info.1,
+                            });
+                            self.update(info.2, unconfirmed_pool);
+                        } else {
+                            self.update(info.2, None)
+                        }
 
-                    self.send_update(addr.clone());
-                    self.save_update(file.clone());
+                        self.send_update(addr.clone());
+                        self.save_update(file.clone());
+                    }
                 }
                 Ok(res) => {
                     warn!("[monitor] unexpected message {:?}", res);